refactor: move `SourceType` inside `rome_js_syntax`
