+++ crates/rome_cli/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 rome_js_formatter = { path = "../rome_js_formatter" }
 rome_formatter = { path = "../rome_formatter" }
 rome_js_parser = { path = "../rome_js_parser" }
+rome_js_syntax = { path = "../rome_js_syntax" }
 rome_diagnostics = { path = "../rome_diagnostics" }
 rome_core = { path = "../rome_core" }
 rome_flags = { path = "../rome_flags" }
