+++ crates/rome_js_parser/src/syntax/program.rs
@@ -4,8 +4,9 @@ use super::module::parse_module_body;
 use super::stmt::parse_statements;
 use crate::state::{ChangeParserState, EnableStrictMode};
 use crate::syntax::stmt::parse_directives;
-use crate::{CompletedMarker, ModuleKind, Parser};
+use crate::{CompletedMarker, Parser};
 use rome_js_syntax::JsSyntaxKind::*;
+use rome_js_syntax::ModuleKind;
 
 // test_err unterminated_unicode_codepoint
 // let s = "\u{200";
