+++ crates/rome_lsp/src/handlers/formatting.rs
@@ -4,8 +4,8 @@ use anyhow::{bail, Result};
 use rome_diagnostics::file::FileId;
 use rome_formatter::IndentStyle;
 use rome_js_formatter::options::{JsFormatOptions, QuoteStyle};
-use rome_js_parser::{parse, SourceType};
-use rome_js_syntax::{TextRange, TokenAtOffset};
+use rome_js_parser::parse;
+use rome_js_syntax::{SourceType, TextRange, TokenAtOffset};
 use std::str::FromStr;
 use tower_lsp::lsp_types::*;
 use tracing::info;
