+++ website/playground/Cargo.toml
@@ -13,6 +13,7 @@ wasm-bindgen = "0.2.79"
 rome_js_formatter = { path = "../../crates/rome_js_formatter" }
 rome_formatter = { path = "../../crates/rome_formatter" }
 rome_js_parser = { path = "../../crates/rome_js_parser", features = ["serde"] }
+rome_js_syntax = { path = "../../crates/rome_js_syntax" }
 rome_diagnostics = { path = "../../crates/rome_diagnostics" }
 rome_analyze = { path = "../../crates/rome_analyze" }
 rome_console = { path = "../../crates/rome_console" }
