+++ website/playground/src/lib.rs
@@ -7,7 +7,8 @@ use rome_diagnostics::Emitter;
 use rome_formatter::IndentStyle;
 use rome_js_formatter::format_node;
 use rome_js_formatter::options::JsFormatOptions;
-use rome_js_parser::{parse, LanguageVariant, SourceType};
+use rome_js_parser::parse;
+use rome_js_syntax::{LanguageVariant, SourceType};
 use serde_json::json;
 use std::io;
 use wasm_bindgen::prelude::*;
