+++ xtask/bench/src/features/parser.rs
@@ -2,8 +2,8 @@ use crate::BenchmarkSummary;
 use itertools::Itertools;
 use rome_diagnostics::file::SimpleFile;
 use rome_diagnostics::{Diagnostic, Emitter, Severity};
-use rome_js_parser::{parse_common, Parse, SourceType};
-use rome_js_syntax::JsAnyRoot;
+use rome_js_parser::{parse_common, Parse};
+use rome_js_syntax::{JsAnyRoot, SourceType};
 use std::fmt::{Display, Formatter};
 use std::ops::Add;
 use std::time::Duration;
